+++ examples/nlp/text_classification/text_classification_with_bert.py
@@ -98,8 +98,8 @@ You may restore the saved model like this:
 import pytorch_lightning as pl
 from omegaconf import DictConfig
 
-from nemo.collections.nlp.models.nlp_model import NLPDDPPlugin
 from nemo.collections.nlp.models.text_classification import TextClassificationModel
+from nemo.collections.nlp.parts.nlp_overrides import NLPDDPPlugin
 from nemo.core.config import hydra_runner
 from nemo.utils import logging
 from nemo.utils.exp_manager import exp_manager
