+++ examples/nlp/token_classification/token_classification_train.py
@@ -18,7 +18,7 @@ import pytorch_lightning as pl
 from omegaconf import DictConfig, OmegaConf
 
 from nemo.collections.nlp.models import TokenClassificationModel
-from nemo.collections.nlp.models.nlp_model import NLPDDPPlugin
+from nemo.collections.nlp.parts.nlp_overrides import NLPDDPPlugin
 from nemo.core.config import hydra_runner
 from nemo.utils import logging
 from nemo.utils.exp_manager import exp_manager
