+++ nemo/collections/nlp/models/nlp_model.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2020, NVIDIA CORPORATION.  All rights reserved.
+# Copyright (c) 2021, NVIDIA CORPORATION.  All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,23 +19,18 @@ import os
 import shutil
 import tarfile
 import tempfile
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, Optional, Union
 
 import torch
 from megatron import mpu
 from megatron.checkpointing import get_checkpoint_version, set_checkpoint_version
-from megatron.initialize import _set_random_seed
 from omegaconf import DictConfig, OmegaConf
 from pytorch_lightning import Trainer
 from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.core.lightning import LightningModule
-from pytorch_lightning.overrides import LightningDistributedModule
-from pytorch_lightning.plugins.environments.cluster_environment import ClusterEnvironment
-from pytorch_lightning.plugins.training_type.ddp import DDPPlugin
 from pytorch_lightning.trainer.connectors.checkpoint_connector import CheckpointConnector
 from pytorch_lightning.utilities import rank_zero_only, rank_zero_warn
 from pytorch_lightning.utilities.cloud_io import atomic_save
-from torch.nn.parallel import DistributedDataParallel
 from transformers import TRANSFORMERS_CACHE
 
 from nemo.collections.common.tokenizers.huggingface.auto_tokenizer import AutoTokenizer
@@ -522,160 +517,3 @@ class NLPCheckpointConnector(CheckpointConnector):
                     )
                     atomic_save(checkpoint, filepath)
         return None
-
-
-class NLPDDPPlugin(DDPPlugin):
-    """ DDP plugin for Pytorch Lightning. Needed to customize DDP for model parallel models.
-    """
-
-    distributed_backend = "ddp"
-
-    def __init__(
-        self,
-        parallel_devices: Optional[List[torch.device]] = None,
-        num_nodes: int = 1,
-        cluster_environment: ClusterEnvironment = None,
-        sync_batchnorm: bool = False,
-        **kwargs: Union[Any, Dict[str, Any]],
-    ) -> None:
-        super().__init__(parallel_devices, num_nodes, cluster_environment, sync_batchnorm, **kwargs)
-
-    def init_ddp_connection(self, global_rank: int, world_size: int) -> None:
-        # call PTL init ddp
-        super().init_ddp_connection(global_rank, world_size)
-
-        # init model parallel
-        app_state = AppState()
-
-        if app_state.model_parallel_size is not None:
-
-            if isinstance(self.lightning_module.bert_model, MegatronBertEncoder):
-
-                if app_state.model_parallel_group is None:
-                    self.init_model_parallel(app_state.global_rank, app_state.world_size)
-
-    def start_training(self, trainer: 'Trainer') -> None:
-        """ PTL Hook that is called after DPP is initialized. """
-
-        if isinstance(self.lightning_module.bert_model, MegatronBertEncoder):
-            app_state = AppState()
-            if app_state.model_parallel_size is not None:
-                # mpu grad clipping needs parameters to have the attribute model_parallel
-                parameters = self.lightning_module.parameters()
-                for p in parameters:
-                    if not hasattr(p, 'model_parallel'):
-                        p.model_parallel = False
-
-                # TODO: figure out how to override clip gradients again
-                # Update PTL trainer to use our _clip_gradients
-                # self._trainer.accelerator_backend._clip_gradients = self._clip_gradients
-
-                if get_checkpoint_version():
-                    # Restored from .nemo, checkpoint_version will already be set
-                    pass
-                elif trainer.resume_from_checkpoint is not None:
-                    # PTL auto-resuming, need to update checkpoint name
-                    # update path based on model parallel rank
-                    filepath = trainer.resume_from_checkpoint
-                    dirname = os.path.dirname(os.path.dirname(filepath))
-                    basename = os.path.basename(filepath)
-                    filepath = f'{dirname}/mp_rank_{app_state.model_parallel_rank:02d}/{basename}'
-                    trainer.resume_from_checkpoint = filepath
-                    logging.info(f'Resuming training from checkpoint {trainer.resume_from_checkpoint}')
-                    # need to set checkpoint version for megatron-lm
-                    checkpoint_version = torch.load(trainer.resume_from_checkpoint).get('checkpoint_version', None)
-                    if checkpoint_version is not None:
-                        set_checkpoint_version(checkpoint_version)
-                    else:
-                        logging.warning('Megatron-lm checkpoint version not found. Setting checkpoint_version to 0.')
-                        set_checkpoint_version(0)
-                else:
-                    logging.info(
-                        f"Restoring from pretrained model parallel checkpoint: {self.lightning_module.bert_model._restore_path}"
-                    )
-                    self.lightning_module.bert_model.restore_weights(self.lightning_module.bert_model._restore_path)
-
-            self.lightning_module.register_megatron_checkpoint_version()
-
-        return super().start_training(trainer)
-
-    def start_testing(self, trainer: 'Trainer') -> None:
-        """ PTL Hook that is called after DPP is initialized. """
-        app_state = AppState()
-
-        if app_state.model_parallel_size is not None:
-
-            if isinstance(self.lightning_module.bert_model, MegatronBertEncoder):
-                # check megatron checkpoint version
-                checkpoint_version = get_checkpoint_version()
-                if checkpoint_version is None:
-                    raise ValueError("Unable to find megatron checkpoint version.")
-
-        return super().start_testing(trainer)
-
-    def configure_ddp(self):
-        """ Override LightningModule ddp if using model parallel.
-            Sets find_unused_parameters to True.
-        """
-
-        app_state = AppState()
-
-        if app_state.model_parallel_size is not None:
-            logging.info(f"Configuring DDP for model parallelism.")
-
-            # With model parallelism, multiple GPUs form a large "logical GPU"
-            # this means that data parallel groups span multiple GPUs
-            # and are non-trivial
-            device_ids = self.determine_ddp_device_ids()
-            self._model = DistributedDataParallel(
-                LightningDistributedModule(self.model),
-                device_ids=device_ids,
-                output_device=device_ids[0],
-                process_group=app_state.data_parallel_group,
-                **self._ddp_kwargs,
-            )
-
-        else:
-            super().configure_ddp()
-
-    def init_model_parallel(self, global_rank: int, world_size: int) -> None:
-        """ Initializes Megatron-LM model parallel if using model parallelism.
-
-        Args:
-            global_rank (int): the global process index.
-            world_size (int): the total number of GPUs, num_nodes * num_gpus
-            is_slurm_managing_tasks (bool, optional): is the cluster managed by SLURM.
-        """
-        app_state = AppState()
-
-        # we initialize megatron-lm model parallel and data parallel groups
-        # after initializing DDP with PTL.
-        if app_state.model_parallel_size is not None:
-            if torch.distributed.is_initialized():
-                mpu.initialize_model_parallel(app_state.model_parallel_size)
-                app_state.model_parallel_group = mpu.get_model_parallel_group()
-                app_state.data_parallel_group = mpu.get_data_parallel_group()
-                app_state.model_parallel_rank = mpu.get_tensor_model_parallel_rank()
-                app_state.data_parallel_rank = mpu.get_data_parallel_rank()
-                app_state.data_parallel_size = mpu.get_data_parallel_world_size()
-                logging.info(f'mp_rank: {app_state.model_parallel_rank}')
-                logging.info(f'dp_rank: {app_state.data_parallel_rank}')
-                # TODO: get random seed from PTL
-                seed = os.environ.get("PL_GLOBAL_SEED", 1234)
-                # random seed must be set for megatron model parallel init
-                _set_random_seed(seed)
-
-    @property
-    def distributed_sampler_kwargs(self):
-        app_state = AppState()
-        if app_state.model_parallel_size is not None:
-            # When using model parallel, data parallel groups are non-trivial and they
-            # correspond to the logical GPUs. This means that the GPUs that form a
-            # single logical GPU all need to get the same batch of data.
-            distributed_sampler_kwargs = dict(
-                num_replicas=app_state.data_parallel_size, rank=app_state.data_parallel_rank
-            )
-            return distributed_sampler_kwargs
-
-        else:
-            return super().distributed_sampler_kwargs
